pyproject.toml
[project]
name = "mlops-project"
version = "0.1.0"
description = "MLOPS project to aply concepts of it"
authors = [
    {name = "Aldrey Sandre",email = "165036139+Code-AldreySandre@users.noreply.github.com"}
]
license = {text = "MIT License"}
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask==2.2.5",
    "gunicorn==23.0.0",
    "numpy==2.1.3",
    "pandas==2.2.3",
    "PyYAML==6.0.2",
    "scikit-learn==1.6.1",
    "tensorflow==2.19.0"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
